+++ packages/mobx-state-tree/src/core/type/type.ts
@@ -152,17 +152,9 @@ export interface ISimpleType<T> extends IType<T, T, T> {}
 
 export type Primitives = ModelPrimitive | null | undefined
 
-// add the interface to the object, but respect the primitives
-export type TAndInterface<T, I> = (Exclude<T, Primitives> & I) | Extract<T, Primitives>
-
 export interface IComplexType<C, S, T> extends IType<C, S, T> {
     // fake, only used for typing
     readonly "!!complexType": undefined
-
-    create(
-        snapshot?: C,
-        environment?: any
-    ): TAndInterface<T, { toJSON?(): S } & IStateTreeNode<C, S>>
 }
 
 export interface IAnyComplexType extends IComplexType<any, any, any> {}
@@ -173,11 +165,6 @@ export type ExtractT<T extends IAnyType> = T extends IType<any, any, infer X> ? 
 export type ExtractCST<IT extends IAnyType> = IT extends IType<infer C, infer S, infer T>
     ? C | S | T
     : never
-
-export type ExtractIStateTreeNode<C, S, T> =
-    // if the instance is a primitive then keep it as is (it is not a state tree node)
-    // else it is a state tree node, but respect primitives
-    T extends ModelPrimitive ? T : TAndInterface<T, IStateTreeNode<C, S>>
 
 export type Instance<T> = T extends IStateTreeNode
     ? T