+++ packages/mobx-state-tree/src/types/complex-types/array.ts
@@ -26,7 +26,6 @@ import {
     IJsonPatch,
     INode,
     isArray,
-    isMutable,
     isNode,
     isPlainObject,
     isStateTreeNode,
@@ -42,12 +41,30 @@ import {
     OptionalProperty,
     ExtractS,
     ExtractC,
-    ExtractT
+    ExtractT,
+    ExtractCST
 } from "../../internal"
 
 export interface IMSTArray<IT extends IAnyType>
     extends IObservableArray<ExtractT<IT>>,
-        IStateTreeNode<ExtractC<IT>, ExtractS<IT>> {}
+        IStateTreeNode<ExtractC<IT>[] | undefined, ExtractS<IT>[]> {
+    // needs to be split or else it will complain about not being compatible with the array interface
+    push(...items: ExtractT<IT>[]): number
+    push(...items: ExtractCST<IT>[]): number
+
+    concat(...items: ConcatArray<ExtractT<IT>>[]): ExtractT<IT>[]
+    concat(...items: ConcatArray<ExtractCST<IT>>[]): ExtractT<IT>[]
+
+    concat(...items: (ExtractT<IT> | ConcatArray<ExtractT<IT>>)[]): ExtractT<IT>[]
+    concat(...items: (ExtractCST<IT> | ConcatArray<ExtractCST<IT>>)[]): ExtractT<IT>[]
+
+    splice(start: number, deleteCount?: number): ExtractT<IT>[]
+    splice(start: number, deleteCount: number, ...items: ExtractT<IT>[]): ExtractT<IT>[]
+    splice(start: number, deleteCount: number, ...items: ExtractCST<IT>[]): ExtractT<IT>[]
+
+    unshift(...items: ExtractT<IT>[]): number
+    unshift(...items: ExtractCST<IT>[]): number
+}
 
 export interface IArrayType<IT extends IAnyType>
     extends IComplexType<ExtractC<IT>[] | undefined, ExtractS<IT>[], IMSTArray<IT>>,
@@ -114,9 +131,9 @@ export class ArrayType<IT extends IAnyType, C = ExtractC<IT>, S = ExtractS<IT>> 
     }
 
     willChange(change: IArrayWillChange<any> | IArrayWillSplice<any>): Object | null {
-        const node = getStateTreeNode(change.object as IStateTreeNode<C, S>)
+        const node = getStateTreeNode(change.object as IMSTArray<IT>)
         node.assertWritable()
-        const subType = (node.type as ArrayType<any, any, any>).subType
+        const subType = (node.type as ArrayType<IT>).subType
         const childNodes = node.getChildren()
         let nodes = null
 
@@ -175,7 +192,7 @@ export class ArrayType<IT extends IAnyType, C = ExtractC<IT>, S = ExtractS<IT>> 
     }
 
     didChange(this: {}, change: IArrayChange<any> | IArraySplice<any>): void {
-        const node = getStateTreeNode(change.object as IStateTreeNode<C, S>)
+        const node = getStateTreeNode(change.object as IMSTArray<IT>)
         switch (change.type) {
             case "update":
                 return void node.emitPatch(